chore: move reconciler to reconciler package
- Move reconciler to reconciler package
- Move SyncStatusClient to a new syncclient package, to avoid import
  cycles betwen parse and reconciler.
- Make some structs, fields, methods, funcs public to allow use between
  packages.
