+++ pkg/parse/cache_test.go
@@ -1,0 +1,77 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package parse
+
+import (
+	"testing"
+
+	"github.com/google/go-cmp/cmp"
+	"kpt.dev/configsync/pkg/core"
+	"kpt.dev/configsync/pkg/core/k8sobjects"
+	"kpt.dev/configsync/pkg/importer/analyzer/ast"
+	"kpt.dev/configsync/pkg/metadata"
+)
+
+func TestSplitObjects(t *testing.T) {
+	testCases := []struct {
+		name             string
+		objs             []ast.FileObject
+		knownScopeObjs   []ast.FileObject
+		unknownScopeObjs []ast.FileObject
+	}{
+		{
+			name: "no unknown scope objects",
+			objs: []ast.FileObject{
+				k8sobjects.Namespace("namespaces/prod", core.Label("environment", "prod")),
+				k8sobjects.Role(core.Namespace("prod")),
+			},
+			knownScopeObjs: []ast.FileObject{
+				k8sobjects.Namespace("namespaces/prod", core.Label("environment", "prod")),
+				k8sobjects.Role(core.Namespace("prod")),
+			},
+		},
+		{
+			name: "has unknown scope objects",
+			objs: []ast.FileObject{
+				k8sobjects.ClusterRole(
+					core.Annotation(metadata.UnknownScopeAnnotationKey, metadata.UnknownScopeAnnotationValue),
+				),
+				k8sobjects.Namespace("namespaces/prod", core.Label("environment", "prod")),
+				k8sobjects.Role(core.Namespace("prod")),
+			},
+			knownScopeObjs: []ast.FileObject{
+				k8sobjects.Namespace("namespaces/prod", core.Label("environment", "prod")),
+				k8sobjects.Role(core.Namespace("prod")),
+			},
+			unknownScopeObjs: []ast.FileObject{
+				k8sobjects.ClusterRole(
+					core.Annotation(metadata.UnknownScopeAnnotationKey, metadata.UnknownScopeAnnotationValue),
+				),
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			gotKnownScopeObjs, gotUnknownScopeObjs := splitObjects(tc.objs)
+			if diff := cmp.Diff(tc.knownScopeObjs, gotKnownScopeObjs, ast.CompareFileObject); diff != "" {
+				t.Errorf("cmp.Diff(tc.knownScopeObjs, gotKnownScopeObjs) = %v", diff)
+			}
+			if diff := cmp.Diff(tc.unknownScopeObjs, gotUnknownScopeObjs, ast.CompareFileObject); diff != "" {
+				t.Errorf("cmp.Diff(tc.unknownScopeObjs, gotUnknownScopeObjs) = %v", diff)
+			}
+		})
+	}
+}
