+++ pkg/parse/updater.go
@@ -51,10 +51,6 @@ type Updater struct {
 	updateMux sync.RWMutex
 }
 
-func (u *Updater) needToUpdateWatch() bool {
-	return u.Remediator.NeedsUpdate()
-}
-
 // HasManagementConflict returns true when conflict errors have been encountered
 // by the Applier or Remediator for at least one currently managed object.
 func (u *Updater) HasManagementConflict() bool {
@@ -65,11 +61,6 @@ func (u *Updater) HasManagementConflict() bool {
 // Applier or Remediator.
 func (u *Updater) ManagementConflicts() []status.ManagementConflictError {
 	return u.SyncErrorCache.conflictHandler.ConflictErrors()
-}
-
-// Remediating returns true if the Remediator is remediating.
-func (u *Updater) Remediating() bool {
-	return u.Remediator.Remediating()
 }
 
 // Update does the following:
@@ -83,7 +74,7 @@ func (u *Updater) Remediating() bool {
 // Any errors returned will be prepended with any known conflict errors from the
 // remediator. This is required to preserve errors that have been reported by
 // another reconciler.
-func (u *Updater) Update(ctx context.Context, cache *cacheForCommit) status.MultiError {
+func (u *Updater) Update(ctx context.Context, cache *CacheForCommit) status.MultiError {
 	u.updateMux.Lock()
 	defer u.updateMux.Unlock()
 
@@ -92,7 +83,7 @@ func (u *Updater) Update(ctx context.Context, cache *cacheForCommit) status.Mult
 
 // update performs most of the work for `Update`, making it easier to
 // consistently prepend the conflict errors.
-func (u *Updater) update(ctx context.Context, cache *cacheForCommit) status.MultiError {
+func (u *Updater) update(ctx context.Context, cache *CacheForCommit) status.MultiError {
 	// Stop remediator workers.
 	// This prevents objects been updated in the wrong order (dependencies).
 	// Continue watching previously declared objects and updating the queue.
@@ -102,50 +93,50 @@ func (u *Updater) update(ctx context.Context, cache *cacheForCommit) status.Mult
 	// Update the declared resources (source of truth for the Remediator).
 	// After this, any objects removed from the declared resources will no
 	// longer be remediated, if they drift.
-	if !cache.declaredResourcesUpdated {
-		objs := filesystem.AsCoreObjects(cache.objsToApply)
-		_, err := u.declare(ctx, objs, cache.source.commit)
+	if !cache.DeclaredResourcesUpdated {
+		objs := filesystem.AsCoreObjects(cache.ObjsToApply)
+		_, err := u.declare(ctx, objs, cache.Source.Commit)
 		if err != nil {
 			return err
 		}
 		// Add new resources to the watch list, without removing old ones.
 		// This ensures controller conflicts are caught while the applier is running.
 		declaredGVKs, _ := u.Resources.DeclaredGVKs()
-		err = u.addWatches(ctx, declaredGVKs, cache.source.commit)
+		err = u.addWatches(ctx, declaredGVKs, cache.Source.Commit)
 		if err != nil {
 			return err
 		}
 		// Only mark the declared resources as updated if there were no (non-blocking) parse errors.
 		// This ensures the update will be retried until parsing fully succeeds.
-		if cache.parserErrs == nil {
-			cache.declaredResourcesUpdated = true
+		if cache.ParserErrs == nil {
+			cache.DeclaredResourcesUpdated = true
 		}
 	}
 
 	// Apply the declared resources
-	if !cache.applied {
+	if !cache.Applied {
 		declaredObjs, _ := u.Resources.DeclaredObjects()
-		if err := u.apply(ctx, declaredObjs, cache.source.commit); err != nil {
+		if err := u.apply(ctx, declaredObjs, cache.Source.Commit); err != nil {
 			return err
 		}
 		// Only mark the commit as applied if there were no (non-blocking) parse errors.
 		// This ensures the apply will be retried until parsing fully succeeds.
-		if cache.parserErrs == nil {
-			cache.applied = true
+		if cache.ParserErrs == nil {
+			cache.Applied = true
 		}
 	}
 
 	// Update the resource watches (triggers for the Remediator).
-	if !cache.watchesUpdated {
+	if !cache.WatchesUpdated {
 		declaredGVKs, _ := u.Resources.DeclaredGVKs()
-		err := u.updateWatches(ctx, declaredGVKs, cache.source.commit)
+		err := u.updateWatches(ctx, declaredGVKs, cache.Source.Commit)
 		if err != nil {
 			return err
 		}
 		// Only mark the watches as updated if there were no (non-blocking) parse errors.
 		// This ensures the update will be retried until parsing fully succeeds.
-		if cache.parserErrs == nil {
-			cache.watchesUpdated = true
+		if cache.ParserErrs == nil {
+			cache.WatchesUpdated = true
 		}
 	}
