+++ pkg/syncclient/large_request.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package parse
+package syncclient
 
 import (
 	"kpt.dev/configsync/pkg/util"
